Add seed data to Category table
A new migration has been created to add seed data to the Category table. The categories added are "Action", "SciFi", and "History" with respective display orders. Corresponding changes have also been made in the ApplicationDbContextModelSnapshot file to reflect these additions. In case of a rollback, the migration will remove this seeded data from the Categories table.